refactor: 로그아웃 위치 이동
